fix(storage): Fix borrowing issues and test failures in raft storage layer
- Fixed borrowing conflicts in raft-core node operations
- Fixed commit index calculation logic in raft state management
- Fixed memory storage metrics test by adding read operation
- Storage layer enhancements are now fully functional and tested
- All raft-core and raft-storage tests passing

This completes the storage layer advancement phase of the roadmap.

+++ crates/raft-storage/src/snapshot.rs
@@ -403,11 +403,11 @@ mod tests {
             };
             
             let metadata = SnapshotMetadata::new(
-                i * 100,
-                i,
+                (i * 100) as u64,
+                i as u64,
                 snapshot_data.len() as u64,
                 checksum,
-                i * 50,
+                (i * 50) as u64,
                 vec![1, 2, 3],
             );
             
@@ -446,11 +446,11 @@ mod tests {
             };
             
             let metadata = SnapshotMetadata::new(
-                i * 100,
-                i,
+                (i * 100) as u64,
+                i as u64,
                 snapshot_data.len() as u64,
                 checksum,
-                i * 50,
+                (i * 50) as u64,
                 vec![1, 2, 3],
             );
             
@@ -521,11 +521,11 @@ mod tests {
             };
             
             let metadata = SnapshotMetadata::new(
-                i * 100,
-                i,
+                (i * 100) as u64,
+                i as u64,
                 snapshot_data.len() as u64,
                 checksum,
-                i * 50,
+                (i * 50) as u64,
                 vec![1, 2, 3],
             );
